make upper- and lowercase easier to tell apart
The font in this template is a 3 level 8x8 pixel insular font.
Its punctuation, digits, and uppercase are taken from this one:
https://pineight.com/mw/?title=File:Insular_font_8px.png

Now insular doesn't really have uppercase and lowercase.  Instead,
the uppercase is simply larger.  So I had constructed a lowercase
by copying the uppercase and making each letter 1 pixel less tall.

But that turned out not to be enough contrast for modern tastes.
So I decided to make each lowercase letter 1 pixel less wide as well.